Create README.md
# Mymotion-cut-internship-project4
Thrilled and excited to share my 4th week Front-end project at MotionCut. Developed an image gallery in my 4th week of internship at MotionCut. Using HTML,CSS and Javascript created an interactive image gallery
This is the week 4 project which I have made using html,css and javscript.
It is a responsive image gallery which is user friendly, the user can enlarge the image in size and can navigate in between the images